requirements.txt
arrow==0.17.0
beautifulsoup4==4.9.3
certifi==2020.6.20
chardet==3.0.4
cloudscraper==1.2.48
config==0.5.0.post0
html5lib==1.1
idna==2.10
lxml==4.5.2
psycopg2-binary==2.8.6
pyparsing==2.4.7
python-dateutil==2.8.1
requests==2.24.0
requests-toolbelt==0.9.1
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.20
uri==2.0.1
urllib3==1.25.10
urlopen==1.0.0
utils==1.0.1
webencodings==0.5.1
wget==3.2

pytz~=2020.1
PyJWT~=1.7.1